Add python script to gather CityObjectTypes.
The script grabs the CityGML XSD files from schemas.opengis.net
and parses them for elements with types that extend the
`AbstractFeatureType` which is the base class of all CityObjects.

The scipt then grabs the existing CityObjectTypes from the
cityobject.h header. Optionally these can be compared with
the `check` argument.

With the optional `fix` argument the script will update the
cityobject.h file to contain all the elements from the XSD
schema and remove all old elements.